rename: 컨벤션에 맞춰 컴포넌트, 파일명 수정
